+++ app/src/main/java/com/CMPUT301F21T30/Habiteer/ui/UserNotifications/NotificationAdapter.java
@@ -1,4 +1,4 @@
-package com.CMPUT301F21T30.Habiteer;
+package com.CMPUT301F21T30.Habiteer.ui.UserNotifications;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,17 +9,13 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
+import com.CMPUT301F21T30.Habiteer.R;
+import com.CMPUT301F21T30.Habiteer.Session;
 
 import java.util.ArrayList;
 
 public class NotificationAdapter extends RecyclerView.Adapter<NotificationAdapter.ViewHolder> {
     private ArrayList<String> requested;
-    private User currentUser;
     private int selectedIndex = RecyclerView.NO_POSITION;
 
     public NotificationAdapter(ArrayList<String> requested){
@@ -30,29 +26,13 @@ public class NotificationAdapter extends RecyclerView.Adapter<NotificationAdapte
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.requested_content, parent, false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.viewholder_follow_notification, parent, false);
         return new NotificationAdapter.ViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
         holder.userName.setText(requested.get(position));
-
-//        Code for getting user given an email
-//        FirebaseFirestore db = FirebaseFirestore.getInstance();
-//        DocumentReference docRef = db.collection("Users").document(requested).get();
-//        docRef.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
-//            @Override
-//            public void onSuccess(DocumentSnapshot documentSnapshot) {
-//                currentUser = documentSnapshot.toObject(User.class);
-//            }
-//        }).addOnFailureListener(new OnFailureListener() {
-//            @Override
-//            public void onFailure(@NonNull Exception e) {
-//
-//            }
-//        });
-
     }
 
     @Override
